--- source/Sashimi.AzureCloudService.Tests/TaskExtensions.cs
@@ -1,9 +1,0 @@
-﻿using System.Threading.Tasks;
-
-public static class TaskExtensions
-{
-    public static void Ignore(this Task _)
-    {
-
-    }
-}
