+++ source/Sashimi.AzureCloudService/Endpoints/AzureCloudServiceDeploymentTargetTypeProvider.cs
@@ -1,7 +1,6 @@
 using System;
 using System.Collections.Generic;
 using FluentValidation;
-using Octopus.Server.Extensibility.HostServices.Mapping;
 using Sashimi.Server.Contracts;
 using Sashimi.Server.Contracts.ActionHandlers;
 using Sashimi.Server.Contracts.Accounts;
@@ -20,6 +19,7 @@ namespace Sashimi.AzureCloudService.Endpoints
         {
             get { yield return AccountTypes.AzureSubscriptionAccountType; }
         }
+
         public IActionHandler HealthCheckActionHandlerForTargetType()
         {
             return new AzureCloudServiceHealthCheckActionHandler();
