+++ source/Sashimi.Tests.Shared/Extensions/TaskExtensions.cs
@@ -1,0 +1,12 @@
+﻿using System.Threading.Tasks;
+
+namespace Sashimi.Tests.Shared.Extensions
+{
+    public static class TaskExtensions
+    {
+        public static void Ignore(this Task _)
+        {
+
+        }
+    }
+}
