+++ source/Server.Contracts/ActionHandlers/IActionHandler.cs
@@ -1,10 +1,5 @@
 ﻿namespace Sashimi.Server.Contracts.ActionHandlers
 {
-    public interface IActionHandlerWithAccount: IActionHandler
-    {
-        string[] StepBasedVariableNameForAccountIds { get; }
-    }
-
     /// <summary>
     /// Implementors of this interface must not keep state so that they can be reusable between steps and deployments
     /// </summary>
