+++ source/Server.Contracts/ActionHandlers/IActionHandlerWithAccount.cs
@@ -1,0 +1,7 @@
+﻿namespace Sashimi.Server.Contracts.ActionHandlers
+{
+    public interface IActionHandlerWithAccount: IActionHandler
+    {
+        string[] StepBasedVariableNameForAccountIds { get; }
+    }
+}
